tests/requirements.txt
pytest
boto3
requests
python-dotenv
